GitHub: use bitcoind v26.0 for CI, remove default value
To make it less confusing what version of bitcoind is actually
installed, we now require the version to be specified as a command line
argument.

Because we tag the version in the docker image tag as bitcoin-core:XX
but the binary internally is located under /opt/bitcoin-XX.Y/, we
manually set Y to 0.

+++ scripts/install_bitcoind.sh
@@ -2,9 +2,15 @@
 
 set -ev
 
-BITCOIND_VERSION=${BITCOIN_VERSION:-25.0}
+BITCOIND_VERSION=$1
+
+if [ -z "$BITCOIND_VERSION" ]; then
+  echo "Must specify a version of bitcoind to install."
+  echo "Usage: install_bitcoind.sh <version>"
+  exit 1
+fi
 
 docker pull lightninglabs/bitcoin-core:$BITCOIND_VERSION
 CONTAINER_ID=$(docker create lightninglabs/bitcoin-core:$BITCOIND_VERSION)
-sudo docker cp $CONTAINER_ID:/opt/bitcoin-$BITCOIND_VERSION/bin/bitcoind /usr/local/bin/bitcoind
+sudo docker cp $CONTAINER_ID:/opt/bitcoin-$BITCOIND_VERSION.0/bin/bitcoind /usr/local/bin/bitcoind
 docker rm $CONTAINER_ID
